requirements.txt
streamlit
requests
datasets
websocket-client
zstandard